# backend/internal/httpapi/router.go
package httpapi

import (
	"context"
	"encoding/json"
	"errors"
	"io"
	"log"
	"net/http"
	"strings"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	authn "github.com/haru-yoshi-5/learning-web-builder/backend/internal/auth"
	"github.com/haru-yoshi-5/learning-web-builder/backend/internal/concept"
	"github.com/haru-yoshi-5/learning-web-builder/backend/internal/project"
	"github.com/haru-yoshi-5/learning-web-builder/backend/internal/site"
)

type Config struct {
	AllowedOrigins []string
	Authenticator  SessionAuthenticator
	Generator      SiteGenerator
	Advisor        ConceptAdvisor
	Projects       project.Repository
}

type SessionAuthenticator interface {
	Optional(http.Handler) http.Handler
}

type SiteGenerator interface {
	Generate(context.Context, string, concept.Draft) (site.Model, error)
}

// ConceptAdvisor は、生成前のコンセプトを固める相談を1ターン進める。
type ConceptAdvisor interface {
	Chat(context.Context, []concept.Message, concept.Draft) (concept.Reply, error)
}

type generateRequest struct {
	Topic string `json:"topic"`
	// 相談で固めたコンセプト。省略できる。
	// DisallowUnknownFields があるため、フィールドを足すだけで後方互換を保てる。
	Concept concept.Draft `json:"concept"`
}

func NewRouter(config Config) http.Handler {
	router := chi.NewRouter()
	router.Use(middleware.RequestID)
	router.Use(middleware.RealIP)
	router.Use(middleware.Recoverer)
	router.Use(middleware.Timeout(30 * time.Second))
	router.Use(cors(config.AllowedOrigins))

	// AIを呼ぶ経路だけ利用量を絞る。健康チェックや保存まで巻き込むと、
	// 画面が動かなくなる範囲が広がりすぎる。
	aiLimiter := newRateLimiter(aiRequestsPerWindow, aiRateLimitWindow)
	aiSlots := make(chan struct{}, aiConcurrencyLimit)
	limitAI := limitAIUsage(aiLimiter, aiSlots)

	router.Route("/api/v1", func(api chi.Router) {
		api.Get("/health", func(writer http.ResponseWriter, _ *http.Request) {
			writeJSON(writer, http.StatusOK, map[string]string{"service": "learning-web-builder-api", "status": "ok", "version": "0.1.0"})
		})
		api.With(optionalAuthentication(config.Authenticator)).Get("/session", session)
		// 認証は必須にしない。ゲストでも試せることが学習用途では大事なため、
		// 身元が分かるときはユーザー単位、分からないときはIP単位で数える。
		api.With(optionalAuthentication(config.Authenticator), limitAI).Post("/generate", generate(config.Generator))
		api.With(optionalAuthentication(config.Authenticator), limitAI).Post("/concept/chat", conceptChat(config.Advisor))
		api.With(optionalAuthentication(config.Authenticator)).Post("/projects", createProject(config.Projects))
		api.With(optionalAuthentication(config.Authenticator)).Get("/projects", listProjects(config.Projects))
		api.With(optionalAuthentication(config.Authenticator)).Get("/projects/{projectId}", getProject(config.Projects))
		api.With(optionalAuthentication(config.Authenticator)).Put("/projects/{projectId}", updateProject(config.Projects))
		api.With(optionalAuthentication(config.Authenticator)).Post("/projects/{projectId}/quality-results", saveQualityResults(config.Projects))
		api.With(optionalAuthentication(config.Authenticator)).Get("/projects/{projectId}/quality-results", listQualityResults(config.Projects))
	})

	return router
}

func optionalAuthentication(authenticator SessionAuthenticator) func(http.Handler) http.Handler {
	if authenticator == nil {
		return func(next http.Handler) http.Handler {
			return next
		}
	}
	return authenticator.Optional
}

func session(writer http.ResponseWriter, request *http.Request) {
	identity, authenticated := authn.IdentityFromContext(request.Context())
	if !authenticated {
		writeJSON(writer, http.StatusOK, map[string]any{"authenticated": false, "mode": "guest"})
		return
	}
	writeJSON(writer, http.StatusOK, map[string]any{
		"authenticated": true,
		"mode":          "clerk",
		"userId":        identity.UserID,
	})
}

func generate(generator SiteGenerator) http.HandlerFunc {
	return func(writer http.ResponseWriter, request *http.Request) {
		request.Body = http.MaxBytesReader(writer, request.Body, 16<<10)
		var input generateRequest
		decoder := json.NewDecoder(request.Body)
		decoder.DisallowUnknownFields()
		if err := decoder.Decode(&input); err != nil {
			writeJSON(writer, http.StatusBadRequest, map[string]string{"error": "invalid request body"})
			return
		}
		if err := decoder.Decode(&struct{}{}); !errors.Is(err, io.EOF) {
			writeJSON(writer, http.StatusBadRequest, map[string]string{"error": "invalid request body"})
			return
		}
		input.Topic = strings.TrimSpace(input.Topic)
		if input.Topic == "" || len([]rune(input.Topic)) > 100 {
			writeJSON(writer, http.StatusBadRequest, map[string]string{"error": "topic must be between 1 and 100 characters"})
			return
		}
		if err := concept.ValidateDraft(input.Concept); err != nil {
			writeJSON(writer, http.StatusBadRequest, map[string]string{"error": err.Error()})
			return
		}
		// 題材の入力元を1つに決める。concept.topic が食い違っていても
		// 生成には topic しか使わないため、どちらが正なのか分からなくなる。
		input.Concept.Topic = input.Topic

		if generator != nil {
			generatedSite, err := generator.Generate(request.Context(), input.Topic, input.Concept)
			if err == nil {
				err = site.Validate(generatedSite)
			}
			if err == nil {
				writeJSON(writer, http.StatusOK, map[string]any{"site": generatedSite, "provider": "gemini"})
				return
			}
			log.Printf("Gemini generation failed; using static fallback request_id=%s error=%v", middleware.GetReqID(request.Context()), err)
		}

		fallback := site.Sample(input.Topic)
		if err := site.Validate(fallback); err != nil {
			log.Printf("static fallback validation failed request_id=%s error=%v", middleware.GetReqID(request.Context()), err)
			writeJSON(writer, http.StatusInternalServerError, map[string]string{"error": "site generation failed"})
			return
		}
		writeJSON(writer, http.StatusOK, map[string]any{"site": fallback, "provider": "static-sample"})
	}
}

func writeJSON(writer http.ResponseWriter, status int, value any) {
	writer.Header().Set("Content-Type", "application/json; charset=utf-8")
	writer.WriteHeader(status)
	_ = json.NewEncoder(writer).Encode(value)
}