# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Learning Web Builder API
  version: 0.1.0
servers:
  - url: http://localhost:8080/api/v1
paths:
  /health:
    get:
      operationId: getHealth
      responses:
        "200":
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                required: [service, status, version]
                properties:
                  service: { type: string }
                  status: { type: string, const: ok }
                  version: { type: string }
  /session:
    get:
      operationId: getSession
      security:
        - {}
        - clerkBearer: []
      responses:
        "200":
          description: Current Clerk or guest session
          content:
            application/json:
              schema:
                type: object
                required: [authenticated, mode]
                properties:
                  authenticated: { type: boolean }
                  mode: { type: string, enum: [guest, clerk] }
                  userId: { type: string, minLength: 1 }
        "401":
          description: Invalid or expired authentication token
  /generate:
    post:
      operationId: generateSite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [topic]
              properties:
                topic: { type: string, minLength: 1, maxLength: 100 }
                concept:
                  # topicはリクエスト直下の値が正となり、ここに書いた値は上書きされる。
                  $ref: "#/components/schemas/ConceptDraft"
      responses:
        "200":
          description: Generated or fallback site model
          content:
            application/json:
              schema:
                type: object
                required: [site, provider]
                properties:
                  provider: { type: string, enum: [gemini, static-sample] }
                  site: { $ref: "#/components/schemas/SiteModel" }
        "400":
          description: Invalid input
        "429":
          description: 利用量の上限を超えた（Retry-Afterに待ち時間を返す）
        "503":
          description: 同時実行の上限に達した（Retry-Afterに待ち時間を返す）
  /concept/chat:
    post:
      operationId: conceptChat
      summary: 生成前のコンセプトを対話で固める
      description: >-
        会話履歴はサーバーに保存しない。呼び出す側が毎回まとめて送る。
        Geminiを利用できないときは、静的な質問テンプレートを返して画面を止めない。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [messages]
              properties:
                messages:
                  type: array
                  minItems: 1
                  maxItems: 24
                  items:
                    type: object
                    additionalProperties: false
                    required: [role, text]
                    properties:
                      role: { type: string, enum: [user, model] }
                      text: { type: string, minLength: 1, maxLength: 500 }
                draft:
                  $ref: "#/components/schemas/ConceptDraft"
      responses:
        "200":
          description: 相談1ターン分の応答
          content:
            application/json:
              schema:
                type: object
                required: [reply, choices, draft, missing, ready, provider]
                properties:
                  reply: { type: string, minLength: 1, maxLength: 400 }
                  choices:
                    type: array
                    maxItems: 4
                    items: { type: string, maxLength: 40 }
                  draft:
                    $ref: "#/components/schemas/ConceptDraft"
                  missing:
                    type: array
                    description: まだ決まっていない必須項目。サーバーがdraftから導出する。
                    items: { type: string, enum: [topic, audience, goal] }
                  ready:
                    type: boolean
                    description: 必須3項目がそろい、生成へ進めるか。モデルの自己申告ではなくサーバーが判定する。
                  provider: { type: string, enum: [gemini, static-sample] }
        "400":
          description: Invalid input
        "429":
          description: 利用量の上限を超えた（Retry-Afterに待ち時間を返す）
        "503":
          description: 同時実行の上限に達した（Retry-Afterに待ち時間を返す）
  /projects:
    get:
      operationId: listProjects
      security:
        - clerkBearer: []
      responses:
        "200":
          description: Projects owned by the authenticated user, ordered by most recently updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectList"
        "401":
          description: Authentication is required
        "503":
          description: Project persistence is not configured
    post:
      operationId: createProject
      security:
        - clerkBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveProjectRequest"
      responses:
        "201":
          description: Project was saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Invalid request or SiteModel
        "401":
          description: Authentication is required
        "503":
          description: Project persistence is not configured
  /projects/{projectId}:
    get:
      operationId: getProject
      security:
        - clerkBearer: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Project owned by the authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Invalid project ID
        "401":
          description: Authentication is required
        "404":
          description: Project was not found for the authenticated owner
        "503":
          description: Project persistence is not configured
    put:
      operationId: updateProject
      security:
        - clerkBearer: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveProjectRequest"
      responses:
        "200":
          description: Project was updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Invalid request, project ID, or SiteModel
        "401":
          description: Authentication is required
        "404":
          description: Project was not found for the authenticated owner
        "503":
          description: Project persistence is not configured
  /projects/{projectId}/quality-results:
    get:
      operationId: listQualityResults
      security:
        - clerkBearer: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Quality check history ordered by most recently checked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QualityResultList"
        "400":
          description: Invalid project ID
        "401":
          description: Authentication is required
        "404":
          description: Project was not found for the authenticated owner
        "503":
          description: Project persistence is not configured
    post:
      operationId: saveQualityResults
      security:
        - clerkBearer: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveQualityResultsRequest"
      responses:
        "201":
          description: Quality check snapshot was saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QualityResultList"
        "400":
          description: Invalid project ID or quality results
        "401":
          description: Authentication is required
        "404":
          description: Project was not found for the authenticated owner
        "503":
          description: Project persistence is not configured
components:
  securitySchemes:
    clerkBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ConceptDraft:
      type: object
      additionalProperties: false
      description: >-
        生成前の相談で固めたコンセプト。まだ決めていない項目は空のままにする。
        topic・audience・goalの3つがそろうと生成へ進める。
      properties:
        topic: { type: string, maxLength: 100 }
        audience: { type: string, maxLength: 80 }
        goal: { type: string, maxLength: 120 }
        tone: { type: string, maxLength: 40 }
        mustInclude:
          type: array
          maxItems: 5
          items: { type: string, maxLength: 60 }
    SaveProjectRequest:
      type: object
      additionalProperties: false
      required: [site]
      properties:
        site:
          $ref: "#/components/schemas/SiteModel"
    Project:
      type: object
      additionalProperties: false
      required: [id, site, version, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        site:
          $ref: "#/components/schemas/SiteModel"
        version:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectList:
      type: object
      additionalProperties: false
      required: [projects]
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
    SaveQualityResultsRequest:
      type: object
      additionalProperties: false
      required: [results]
      properties:
        results:
          type: array
          minItems: 1
          maxItems: 4
          items:
            type: object
            additionalProperties: false
            required: [checkKey, passed, detail]
            properties:
              checkKey:
                type: string
                enum: [headings, alt, mobile, axe]
              passed:
                type: boolean
              detail:
                type: string
                minLength: 1
                maxLength: 1000
    QualityResult:
      type: object
      additionalProperties: false
      required: [id, projectId, checkKey, passed, detail, checkedAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        checkKey:
          type: string
          enum: [headings, alt, mobile, axe]
        passed:
          type: boolean
        detail:
          type: string
        checkedAt:
          type: string
          format: date-time
    QualityResultList:
      type: object
      additionalProperties: false
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/QualityResult"
    SiteModel:
      type: object
      required: [id, topic, siteTitle, tagline, theme, sections]
      properties:
        id: { type: string, minLength: 1 }
        topic: { type: string, minLength: 1, maxLength: 100 }
        siteTitle: { type: string, minLength: 1, maxLength: 80 }
        tagline: { type: string, maxLength: 160 }
        theme:
          type: object
          required: [primary, background, text, fontFamily, spacing]
          properties:
            primary: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
            background: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
            text: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
            # 見出し(h2)の色。任意項目で、未指定ならクライアントがprimaryを使う。
            heading: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
            fontFamily: { type: string, enum: [sans, serif, rounded] }
            spacing: { type: integer, minimum: 2, maximum: 10 }
        sections:
          type: array
          minItems: 2
          maxItems: 8
          items:
            type: object
            required: [id, kind, title, body, imageAlt, visible]
            properties:
              id: { type: string, minLength: 1 }
              kind: { type: string, enum: [hero, about, features, gallery, contact] }
              title: { type: string, minLength: 1, maxLength: 80 }
              body: { type: string, maxLength: 800 }
              imageAlt: { type: string, maxLength: 160 }
              visible: { type: boolean }